[Draw2D] Extract custom tool-tip handling from Draw2D Figure
The CustomToolTipManager is only used in the context of palette figures
and should therefore not be part of the generic Draw2D Figure.
